day3: refactor
Use complex to store coordinates instead of a namedtuple.

Switch to dataclass for Part and add new Symbol dataclass.

Switch Graph to hold a single list of symbols and a dictionary of parts
indexed by each part's span.

Invert the logic for finding intersections. Instead of enumerating each
part's border and looking for an intersecting symbol, we now enumerate
each symbol's border and look for an intersecting part. This is simpler
and more efficient knowing part 2 of the day's puzzle.