Automate dayNN creation
Add Makefile `day%` rule which creates that day's puzzle directory,
seeding it with a template `dayNN.py` solution script and using `aoc`
to download the puzzle description and sample input. The rule is
implemented by a new `bin/download` shell script.

Normalize the existing `dayNN.py` scripts to use `INPUT`, `EXAMPLE1`,
and `EXAMPLE2` along with their corresponding text files. If the two
examples are the same, we symlink the second to the first.